+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Display no roll difficulty for non-private rolls without a set difficulty
+- Prevent creation of adversary templates on actors that already have a template
 
 ### Fixed